Clean up
- Use dotenv for API tokens
- Create json directory if it doesn't exist and gitignored
