+++ .env.example
@@ -1,0 +1,2 @@
+NOTION_API_KEY=
+DISCORD_BOT_KEY=
