+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "@notionhq/client": "1.0.4",
+    "dotenv": "16.0.3",
     "node-fetch": "2.6.7"
   }
 }
