--- src/json/MEMBERS_DISCORD.json
@@ -1,1 +1,0 @@
-[]
