--- src/json/MEMBERS_NOTION.json
@@ -1,1 +1,0 @@
-[]
